feat: add loader implementation [WIP]
