+++ packages/next-public/package.json
@@ -20,9 +20,10 @@
   },
   "devDependencies": {
     "@types/node": "^18.19.123",
+    "@types/webpack": "^5.28.5",
     "tsup": "^8.0.2",
     "typescript": "^5.2.2",
-    "webpack": "^5.89.0"
+    "webpack": "^5.101.3"
   },
   "repository": {
     "type": "git",
