+++ packages/next-public/src/types.ts
@@ -1,0 +1,13 @@
+export type PluginOptions = {
+  enabled?: boolean;
+  shouldLog?: boolean;
+} & (
+  | {
+      inputDir: string | string[];
+      outputDir: string;
+      autoDetect?: false;
+    }
+  | {
+      autoDetect: true;
+    }
+);
